zstd compressor refactor
