+++ source/common/compression/zstd/common/base.h
@@ -7,7 +7,6 @@
 #include "zstd.h"
 
 namespace Envoy {
-namespace Extensions {
 namespace Compression {
 namespace Zstd {
 namespace Common {
@@ -29,5 +28,4 @@ protected:
 } // namespace Common
 } // namespace Zstd
 } // namespace Compression
-} // namespace Extensions
 } // namespace Envoy
