+++ source/extensions/compression/zstd/compressor/zstd_compressor_impl.h
@@ -2,8 +2,9 @@
 
 #include "envoy/compression/compressor/compressor.h"
 
-#include "source/extensions/compression/zstd/common/base.h"
-#include "source/extensions/compression/zstd/common/dictionary_manager.h"
+#include "source/common/compression/zstd/common/base.h"
+#include "source/common/compression/zstd/common/dictionary_manager.h"
+#include "source/common/compression/zstd/compressor/zstd_compressor_impl_base.h"
 
 namespace Envoy {
 namespace Extensions {
@@ -12,28 +13,23 @@ namespace Zstd {
 namespace Compressor {
 
 using ZstdCDictManager =
-    Common::DictionaryManager<ZSTD_CDict, ZSTD_freeCDict, ZSTD_getDictID_fromCDict>;
+    Envoy::Compression::Zstd::Common::DictionaryManager<ZSTD_CDict, ZSTD_freeCDict, ZSTD_getDictID_fromCDict>;
 using ZstdCDictManagerPtr = std::unique_ptr<ZstdCDictManager>;
 
 /**
  * Implementation of compressor's interface.
  */
-class ZstdCompressorImpl : public Common::Base,
-                           public Envoy::Compression::Compressor::Compressor,
-                           NonCopyable {
+class ZstdCompressorImpl : public Envoy::Compression::Zstd::Compressor::ZstdCompressorImplBase {
 public:
   ZstdCompressorImpl(uint32_t compression_level, bool enable_checksum, uint32_t strategy,
-                     const ZstdCDictManagerPtr& cdict_manager, uint32_t chunk_size);
-
-  // Compression::Compressor::Compressor
-  void compress(Buffer::Instance& buffer, Envoy::Compression::Compressor::State state) override;
+                     const ZstdCDictManagerPtr& cdict_manager, uint32_t chunk_size):ZstdCompressorImplBase(compression_level,enable_checksum,strategy,cdict_manager,chunk_size){}
 
 private:
-  void process(Buffer::Instance& output_buffer, ZSTD_EndDirective mode);
+  void compressPreprocess() override;
 
-  std::unique_ptr<ZSTD_CCtx, decltype(&ZSTD_freeCCtx)> cctx_;
-  const ZstdCDictManagerPtr& cdict_manager_;
-  const uint32_t compression_level_;
+  void compressProcess(const Buffer::RawSlice& input_slice, Buffer::Instance& accumulation_buffer) override;
+
+  void compressPostprocess() override;
 };
 
 } // namespace Compressor
