+++ source/extensions/extensions_build_config.bzl
@@ -35,7 +35,7 @@ EXTENSIONS = {
     "envoy.compression.brotli.compressor":              "//source/extensions/compression/brotli/compressor:config",
     "envoy.compression.brotli.decompressor":            "//source/extensions/compression/brotli/decompressor:config",
     "envoy.compression.zstd.compressor":                "//source/extensions/compression/zstd/compressor:config",
-    "envoy.compression.zstd.decompressor":              "//source/extensions/compression/zstd/decompressor:config",
+    # "envoy.compression.zstd.decompressor":              "//source/extensions/compression/zstd/decompressor:config",
 
     #
     # Config validators
